[#16] Cleanup & updating to latest packages
[#16] Refactor index route token exchange

Moved index route `afterModel` hook that exchanged the OAuth token for
an AI token to the Auth service.

[#16] Fixed race condition

There was a race condition for users logging in, ember-simple-auth would redirect them to the  route as soon as the first  was invoked. Then the  route would request  before their social token was exchanged for an Ai token, and requests to the API failed because the Ai token was invalid. Now the  route will wait on the login page promise until its complete.

[#16] Fix for nav and offline pre-render

* [x] Fixes flash of nav and offline UI while the dom is loading
* [x] request for current user uses JSON-API content type

fixed translation missing